Create Website
This section includes work from CIS Development Project, the subsequent course after Systems Analysis and Design. It focuses on building the system that was designed in the Systems Analysis and Design course. The main focus of this project was constructing a website for the University of Louisville's Office of Research and Innovation Website,